added a generic table base string char translator, plus a new class "stringtranslationtable". replaced old capitalize$ to make use of the table translator, as example of use. made int from short command.repeat - the instantiator stores an address in there, for which short was a bit short.

+++ trunk/src/c_bind.h
@@ -87,7 +87,7 @@ public:
 
 typedef struct {
     int type;
-    short repeat;
+    int repeat;
     short ign;
     union {
         long num;
